Qt: Move Main.cpp to QtHost.cpp
