+++ pcsx2-qt/QtHost.h
@@ -40,12 +40,8 @@ Q_DECLARE_METATYPE(CDVD_SourceType);
 
 namespace QtHost
 {
-	bool Initialize();
-	void Shutdown();
-
 	/// Sets batch mode (exit after game shutdown).
 	bool InBatchMode();
-	void SetBatchMode(bool enabled);
 
 	/// Executes a function on the UI thread.
 	void RunOnUIThread(const std::function<void()>& func, bool block = false);
