+++ pcsx2-qt/pcsx2-qt.vcxproj.filters
@@ -30,7 +30,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="PrecompiledHeader.cpp" />
-    <ClCompile Include="Main.cpp" />
     <ClCompile Include="QtHost.cpp" />
     <ClCompile Include="MainWindow.cpp" />
     <ClCompile Include="$(IntDir)moc_MainWindow.cpp">
